+++ framework/Database/Migrations/2024_01_01_013858_create_user_roles_table.php
@@ -1,0 +1,23 @@
+<?php
+
+use Framework\Database\Database;
+use Framework\Database\Migration;
+
+return new class extends Migration
+{
+    public function up(): void
+    {
+        Database::query(
+            'CREATE TABLE userRoles (' .
+            'id INT auto_increment,' .
+            'userId INT NOT NULL,' .
+            'roleId INT NOT NULL,' .
+            'createdAt TIMESTAMP DEFAULT CURRENT_TIMESTAMP,' .
+            'updatedAt DATETIME DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,' .
+            'PRIMARY KEY (id),' .
+            'CONSTRAINT `fkUserRolesUserId` FOREIGN KEY (userId) REFERENCES users (id) ON DELETE CASCADE,' .
+            'CONSTRAINT `fkUserRolesRole` FOREIGN KEY (roleId) REFERENCES roles (id) ON DELETE CASCADE' .
+            ') ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;'
+        );
+    }
+};