+++ CHANGELOG.md
@@ -39,6 +39,7 @@ Types of changes: `Added`, `Changed`, `Deprecate`, `Removed`, `Fixed`, `Secruity
 - Added functions to check if the user is logged in
 - Added routes for login and logout
 - Added command to create a new user
+- Added table for roles and user-role-assignment
 
 ### Changed
 - Improved the autoloader to work for the new folder structure of version 2